--- src/main/java/com/monorama/iot_server/dto/request/HealthDataSyncRequestDto.java
@@ -1,7 +1,0 @@
-package com.monorama.iot_server.dto.request;
-
-import java.util.List;
-
-public record HealthDataSyncRequestDto(
-        List<HealthDataRequestDto> healthDataList
-) {}
