+++ src/main/java/com/monorama/iot_server/dto/request/healthdata/HDSyncRequestDto.java
@@ -1,0 +1,7 @@
+package com.monorama.iot_server.dto.request.healthdata;
+
+import java.util.List;
+
+public record HDSyncRequestDto(
+        List<HDRequestDto> healthDataList
+) {}
