+++ src/main/java/com/monorama/iot_server/repository/HDRepository.java
@@ -3,5 +3,5 @@ package com.monorama.iot_server.repository;
 import com.monorama.iot_server.domain.HealthData;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface HealthDataRepository extends JpaRepository<HealthData, Long> {
+public interface HDRepository extends JpaRepository<HealthData, Long> {
 }
