+++ src/main/java/com/monorama/iot_server/service/healthdata/HDService.java
@@ -1,11 +1,11 @@
 package com.monorama.iot_server.service.healthdata;
 
 import com.monorama.iot_server.domain.User;
-import com.monorama.iot_server.dto.request.HealthDataRequestDto;
-import com.monorama.iot_server.dto.request.HealthDataSyncRequestDto;
+import com.monorama.iot_server.dto.request.healthdata.HDRequestDto;
+import com.monorama.iot_server.dto.request.healthdata.HDSyncRequestDto;
 import com.monorama.iot_server.exception.CommonException;
 import com.monorama.iot_server.exception.ErrorCode;
-import com.monorama.iot_server.repository.HealthDataRepository;
+import com.monorama.iot_server.repository.HDRepository;
 import com.monorama.iot_server.repository.UserRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
@@ -13,24 +13,24 @@ import org.springframework.transaction.annotation.Transactional;
 
 @Service
 @RequiredArgsConstructor
-public class HealthDataService {
+public class HDService {
 
     private final UserRepository userRepository;
-    private final HealthDataRepository healthDataRepository;
+    private final HDRepository HDRepository;
 
     @Transactional
-    public void saveRealtime(Long userId, HealthDataRequestDto dto) {
+    public void saveRealtime(Long userId, HDRequestDto dto) {
         User user = userRepository.findById(userId)
                 .orElseThrow(() -> new CommonException(ErrorCode.NOT_FOUND_USER));
-        healthDataRepository.save(dto.toEntity(user));
+        HDRepository.save(dto.toEntity(user));
     }
 
 
     @Transactional
-    public void saveSync(Long userId, HealthDataSyncRequestDto request) {
+    public void saveSync(Long userId, HDSyncRequestDto request) {
         User user = userRepository.findById(userId)
                 .orElseThrow(() -> new CommonException(ErrorCode.NOT_FOUND_USER));
-        healthDataRepository.saveAll(
+        HDRepository.saveAll(
                 request.healthDataList().stream()
                         .map(dto -> dto.toEntity(user))
                         .toList()
